Deterministic generation of IDs
Added a new class, DefaultIdGenerator, to generate serial IDs for elements using base36 encoding with an optional prefix. Modified the Application's constructor to accept an instance of an ID generator, defaulting to DefaultIdGenerator if none is provided.

This makes both testing and, possibly, sever-side rendering more feasible, since ids are entirely predictable now, with the option to override how they're generated.